Replaced HashRouter for BrowserRouter
We use BrowserRouter when we need to do a GET request to be sent to the server. In comparison, HashRouter is useful when you have a static file server and cannot configure the server to handle dynamic routes. In our case, we need to use BrowserRouter to know if the user is signed in. This fixes the bug of the page being stuck in the loading screen.